Omrokkerer komponenter og sletter status i saken + fjerner ettersendingspanel
