--- src/components/fullforteSoknader/FullforteSoknader.module.css
@@ -1,25 +1,0 @@
-.fullforteSoknader {
-  border-top: 7px solid #f1f1f1;
-  margin-top: 49px;
-  padding-top: 40px;
-}
-
-.fullforteSoknader:last-child {
-  margin-bottom: 49px;
-}
-
-.fullfortStatus {
-  margin-left: auto;
-}
-
-.fullforteSoknaderDetaljer {
-  display: flex;
-  gap: 10px;
-  padding: 0 2rem;
-}
-
-.fullfortIkon {
-  width: 48px;
-  height: 48px;
-  flex-shrink: 0;
-}
