--- src/components/paabegynteSoknader/PaabegynteSoknader.module.css
@@ -1,21 +1,0 @@
-.paabegyntSoknad {
-  border-top: 7px solid #f1f1f1;
-  margin-top: 49px;
-  padding-top: 40px;
-}
-
-.paabegyntStatus {
-  margin-left: auto;
-}
-
-.paabegyntSoknadDetaljer {
-  display: flex;
-  gap: 10px;
-  padding: 0 2rem;
-}
-
-.paabegyntIkon {
-  width: 48px;
-  height: 48px;
-  flex-shrink: 0;
-}
